move emulator tests into scripts folder
