+++ package.json
@@ -22,8 +22,9 @@
     "mocha": "nyc mocha 'src/test/**/*.{ts,js}'",
     "prepare": "npm run clean && npm run build -- --build tsconfig.publish.json",
     "test": "npm run lint:quiet && npm run test:compile && npm run mocha",
-    "test:compile": "tsc --project tsconfig.compile.json",
     "test:client-integration": "./scripts/client-integration-tests/run.sh",
+    "test:compile": "tsc --project tsconfig.compile.json",
+    "test:emulator": "./scripts/emulator-tests/run.sh",
     "test:extensions-emulator": "./scripts/extensions-emulator-tests/run.sh",
     "test:hosting": "./scripts/hosting-tests/run.sh",
     "test:triggers-end-to-end": "./scripts/triggers-end-to-end-tests/run.sh"
